bordel: refactoring for better abstraction
Currently with the way bordel is persisting configuration, it is
populating explicit paths. This breaks when configuring a build in the
host and then attempting to do a build from within a container.

To address this, directories that are shareable across build are now
symlinked into the build directory. This means they are now all
available under the bitbake TOPDIR path and thus in the generated OE
configurations the paths can now be based on TOPDIR variable. With this
change we also dropped the need for the .putain config persistence file
and set the variables explicitly during start of bordel script. Finally
the BUILD_ID has been evolved that it can be set on commandline, check
for the existence of a symlink "build" and checks if it can derive the
build id from the link source, or it will default to YYMMDD.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>